chore: Update favicon.ico link, remove unused code, and add zeta.svg and page.tsx
This commit updates the favicon.ico link in layout.tsx, removes the unused pull_request trigger from the deploy workflow, and adds zeta.svg and updates page.tsx in the my_components folder. It also includes updates to globals.css and layout.tsx. The Vercel deployment command is updated to include the "--token=$VERCEL_TOKEN" flag for authentication and the "--yes" flag for automatic confirmation. This commit is part of the ongoing development and deployment process, integrating GitHub Actions and Vercel for continuous deployment. (#1)